requirements.txt
openai==0.10.0
python-dotenv==0.20.0